refactor: move op state registration to workers (#7696)
